requirements/local.txt
-r base.txt

dj-database-url==0.3.0
django-extensions
pep8==1.5.7
pylint
codecov<2.1,>=2.0.5
coverage<4.4,>=4.3.4